:truck: Move ContributorsTable
