+++ frontend/src/components/PaymentTable/Headers.tsx
@@ -5,7 +5,7 @@ import TimeLine from "src/icons/TimeLine";
 import Folder3Line from "src/icons/Folder3Line";
 import MoneyDollarCircleLine from "src/icons/MoneyDollarCircleLine";
 import FocusLine from "src/icons/FocusLine";
-import SortingArrow from "../ContributorsTable/SortingArrow";
+import SortingArrow from "../../pages/ProjectDetails/Contributors/ContributorsTable/SortingArrow";
 import { Field, Sorting } from "src/hooks/usePaymentSorting";
 
 type Props = {
