+++ frontend/src/pages/ProjectDetails/Contributors/ContributorsTable/ContributorsTable.stories.tsx
