+++ frontend/src/pages/ProjectDetails/Contributors/ContributorsTable/View.tsx
@@ -1,4 +1,4 @@
-import Table from "../Table";
+import Table from "../../../../components/Table";
 import { rates } from "src/hooks/useWorkEstimation";
 import { useMemo, useState } from "react";
 import { sortBy } from "lodash";
