+++ frontend/src/pages/ProjectDetails/Contributors/ContributorsTable/index.tsx
